modified:   include/subsystems/drive.hpp 	modified:   src/auton.cpp 	modified:   src/definitions/autons.cpp 	modified:   src/definitions/drive.cpp 	modified:   src/definitions/globals.cpp 	modified:   src/init.cpp 	modified:   src/opcontrol.cpp 	modified:   src/subsystems/latch.cpp 	modified:   src/subsystems/lift.cpp

+++ include/subsystems/drive.hpp
@@ -94,7 +94,7 @@ void resetDriveEncoders();
 /*
 * Gets the avg drive pos
 */
-double avgDriveEncoderValue();
+int avgDriveEncoderValue();
 
 /*
 * Get left position
